Moved some functionality from H2OContext to SupportedDataset in converters package, as requested. Uncommented a call in RDDHandler to ensure using ProductBuilder, as requested.
Several warnings from IntelliJ fixed too.
