+++ core/src/main/scala/water/api/RDDs/RDDsHandler.scala
@@ -17,6 +17,7 @@
 package water.api.RDDs
 
 import org.apache.spark.SparkContext
+import org.apache.spark.h2o.converters.SupportedDataset.ProductFrameBuilder
 import org.apache.spark.h2o.{H2OContext, H2OFrame}
 import org.apache.spark.mllib.regression.LabeledPoint
 import org.apache.spark.rdd.RDD
@@ -63,8 +64,7 @@ class RDDsHandler(val sc: SparkContext, val h2oContext: H2OContext) extends Hand
         case t if t.isInstanceOf[Float] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[Float]], name)
         case t if t.isInstanceOf[Long] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[Long]], name)
         case t if t.isInstanceOf[java.sql.Timestamp] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[java.sql.Timestamp]], name)
-//        case t if t.isInstanceOf[Product] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[Product]], name)
-        case t if t.isInstanceOf[Product] => H2OContext.FromPureProduct(sc, rdd.asInstanceOf[RDD[Product]], name).withDefaultFieldNames()
+        case t if t.isInstanceOf[Product] => ProductFrameBuilder(sc, rdd.asInstanceOf[RDD[Product]], name).withDefaultFieldNames()
         case t => throw new IllegalArgumentException(s"Do not understand type $t")
       }
     }
